Fixing Ticket Number Generating while add new Booking

+++ back end/src/main/java/com/blibli/repository/booking/BookingRepositoryCustom.java
@@ -2,6 +2,7 @@ package com.blibli.repository.booking;
 
 import com.blibli.model.Booking;
 
+import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -11,6 +12,6 @@ public interface BookingRepositoryCustom {
     List<Booking> showById(String id);
     Booking deleteBooking(String id);
     List<Booking> showByEmployeeId(String empId);
-    Integer countBooking();
+    BigInteger countBooking();
     Booking getBookingByTicket(String ticket);
 }
